requirements.txt
django=2.2.3
djangorestframework=3.10.1
django-rest-auth==0.9.5
django-allauth==0.39.1
django-cors-headers==3.0.2